requirements.txt
python-telegram-bot==20.7
sqlalchemy==1.4.31
psycopg2 ==2.9.10
python-dotenv==0.19.2
google-api-python-client==2.37.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
pandas==2.1.3
gspread==6.1.4
oauth2client==4.1.3
google-auth
google-auth-oauthlib
google-api-python-client
protobuf~=5.28.3
alembic==1.7.7
urllib3==2.2.2